Move view_roots to the only file where it is used
